feat: move "layout-with-side-panel" to "Layout"
It is now the de-facto layout that we use.
