+++ services/orchest-webserver/client/src/components/Layout/LayoutWithSidePanel.tsx
@@ -1,7 +1,7 @@
 import Stack, { StackProps } from "@mui/material/Stack";
 import React from "react";
-import { Layout } from "../Layout";
 import { CenteredStack } from "./CenteredStack";
+import { Layout } from "./Layout";
 import { MainContainer } from "./MainContainer";
 import { MainSidePanel } from "./MainSidePanel";
 import { StickyHeader } from "./StickyHeader";
@@ -13,6 +13,13 @@ export type LayoutWithSidePanelProps = StackProps & {
   header?: React.ReactNode;
 };
 
+/**
+ * Provides the share base layout for many views.
+ * This components adds the main menus and toolbars around the view (from `Layout`).
+ *
+ * For convenience you may specify a (sticky) `header` and a `sidePanel`
+ * which is used in many views.
+ */
 export const LayoutWithSidePanel = React.forwardRef<
   HTMLDivElement,
   LayoutWithSidePanelProps
