+++ services/orchest-webserver/client/src/components/Layout/MainSidePanel.tsx
@@ -1,4 +1,5 @@
 import { ResizablePane } from "@/components/ResizablePane";
+import { SxProps, Theme } from "@mui/material";
 import React from "react";
 import {
   MAX_WIDTH,
@@ -6,8 +7,16 @@ import {
   useMainSidePanelWidth,
 } from "./stores/useLayoutStore";
 
-type MainSidePanelProps = { children: React.ReactNode };
-export const MainSidePanel = ({ children }: MainSidePanelProps) => {
+const paneSx: SxProps<Theme> = {
+  backgroundColor: (theme) => theme.palette.grey[100],
+  borderRight: (theme) => `1px solid ${theme.borderColor}`,
+  display: "flex",
+  flexDirection: "column",
+  height: "100%",
+};
+
+/** A resizable & scrollable side pane, commonly used for menus. */
+export const MainSidePanel: React.FC = ({ children }) => {
   const [mainSidePanelWidth, setMainSidePanelWidth] = useMainSidePanelWidth();
 
   return (
@@ -17,13 +26,7 @@ export const MainSidePanel = ({ children }: MainSidePanelProps) => {
       minWidth={MIN_MAIN_SIDE_PANEL_WIDTH}
       maxWidth={MAX_WIDTH}
       position="relative"
-      sx={{
-        backgroundColor: (theme) => theme.palette.grey[100],
-        borderRight: (theme) => `1px solid ${theme.borderColor}`,
-        display: "flex",
-        flexDirection: "column",
-        height: "100%",
-      }}
+      sx={paneSx}
       onSetSize={setMainSidePanelWidth}
     >
       {children}
