+++ services/orchest-webserver/client/src/components/Layout/StatusIcon.tsx
