+++ services/orchest-webserver/client/src/environments-view/BuildStatusIcon.tsx
@@ -1,7 +1,4 @@
-import {
-  StatusIcon,
-  StatusIconStatus,
-} from "@/components/Layout/layout-with-side-panel/StatusIcon";
+import { StatusIcon, StatusIconStatus } from "@/components/Layout/StatusIcon";
 import { EnvironmentImageBuild } from "@/types";
 import React from "react";
 import { isEnvironmentBuilding, isEnvironmentFailedToBuild } from "./common";
