+++ services/orchest-webserver/client/src/environments-view/EnvironmentMenuItem.tsx
@@ -1,4 +1,4 @@
-import { SidePanelMenuItem } from "@/components/Layout/layout-with-side-panel/SidePanelMenuItem";
+import { SidePanelMenuItem } from "@/components/Layout/SidePanelMenuItem";
 import { TStatus } from "@/components/Status";
 import { EnvironmentState } from "@/types";
 import React from "react";
