+++ services/orchest-webserver/client/src/jobs-view/JobMenuItem.tsx
@@ -1,4 +1,4 @@
-import { SidePanelMenuItem } from "@/components/Layout/layout-with-side-panel/SidePanelMenuItem";
+import { SidePanelMenuItem } from "@/components/Layout/SidePanelMenuItem";
 import { JobStatus } from "@/types";
 import React from "react";
 import { JobStatusIcon } from "./JobStatusIcon";
